Prototyped offline embedding functions

+++ src/bioterms/vocabulary/__init__.py
@@ -11,6 +11,7 @@ from bioterms.etc.enums import ConceptPrefix
 from bioterms.etc.utils import check_files_exist
 from bioterms.database import Cache, DocumentDatabase, GraphDatabase, VectorDatabase, get_active_cache, \
     get_active_doc_db, get_active_graph_db, get_active_vector_db
+from bioterms.embedding import ConceptTransformer, EmbeddingContainerV1, EmbeddingContainerFileV1
 from .utils import ALL_VOCABULARIES, get_vocabulary_module, get_vocabulary_status
 
 
@@ -231,6 +232,7 @@ async def load_vocabulary(prefix: ConceptPrefix,
 
 async def embed_vocabulary(prefix: ConceptPrefix,
                            drop_existing: bool = True,
+                           offline: bool = False,
                            doc_db: DocumentDatabase = None,
                            graph_db: GraphDatabase = None,
                            vector_db: VectorDatabase = None,
@@ -239,44 +241,70 @@ async def embed_vocabulary(prefix: ConceptPrefix,
     Embed the vocabulary specified by the prefix.
     :param prefix: The prefix of the vocabulary to embed.
     :param drop_existing: Whether to drop existing embeddings before embedding.
+    :param offline: Whether to operate in offline mode (skip writing to vector database).
     :param doc_db: The document database instance.
     :param graph_db: The graph database instance.
     :param vector_db: The vector database instance.
     """
-    if doc_db is None:
-        doc_db = await get_active_doc_db()
-    if vector_db is None:
-        vector_db = get_active_vector_db()
-
     config = get_vocabulary_config(prefix)
-    status = await get_vocabulary_status(
-        prefix=prefix,
-        doc_db=doc_db,
-        graph_db=graph_db,
-    )
 
-    if not status.loaded:
-        raise RuntimeError(f'Vocabulary {prefix} is not loaded. Cannot embed.')
+    if not offline:
+        if doc_db is None:
+            doc_db = await get_active_doc_db()
+        if vector_db is None:
+            vector_db = get_active_vector_db()
 
-    if drop_existing:
-        await vector_db.delete_vectors_for_prefix(prefix=prefix)
+        status = await get_vocabulary_status(
+            prefix=prefix,
+            doc_db=doc_db,
+            graph_db=graph_db,
+        )
 
-    concept_iter = doc_db.get_terms_iter(
-        prefix=prefix,
-        model_class=config['conceptClass'],
-    )
+        if not status.loaded:
+            raise RuntimeError(f'Vocabulary {prefix} is not loaded. Cannot embed.')
 
-    id_map = await vector_db.insert_concepts(
-        concepts=concept_iter,
-        prefix=prefix,
-        total_concepts=status.concept_count,
-    )
+        if drop_existing:
+            await vector_db.delete_vectors_for_prefix(prefix=prefix)
 
-    await doc_db.update_vector_mapping(
-        prefix=prefix,
-        mapping=id_map,
-    )
+        concept_iter = doc_db.get_terms_iter(
+            prefix=prefix,
+            model_class=config['conceptClass'],
+        )
 
+        id_map = await vector_db.insert_concepts(
+            concepts=concept_iter,
+            prefix=prefix,
+            total_concepts=status.concept_count,
+        )
+
+        await doc_db.update_vector_mapping(
+            prefix=prefix,
+            mapping=id_map,
+        )
+    else:
+        offline_concept_path = os.path.join(CONFIG.data_dir, 'offline', f'{prefix.value}.doc.dump')
+        if not os.path.exists(offline_concept_path):
+            raise ValueError(f'Offline concept file for {prefix} not found at {offline_concept_path}.')
+        offline_embedding_path = os.path.join(CONFIG.data_dir, 'offline', f'{prefix.value}.embed.dump')
+
+        async def concept_iter():
+            async with aiofiles.open(offline_concept_path) as f:
+                async for line in f:
+                    yield config['conceptClass'].model_validate_json(line.strip())
+
+        transformer = ConceptTransformer()
+        async def embed_iter():
+            async for batch in transformer.embed_concepts(
+                concepts=concept_iter(),
+            ):
+                for concept_id, vector in batch:
+                    yield EmbeddingContainerV1(
+                        concept_id=concept_id,
+                        vector=vector,
+                    )
+
+        embedding_file = EmbeddingContainerFileV1(offline_embedding_path)
+        await embedding_file.write(embed_iter())
 
 def get_vocabulary_license(prefix: ConceptPrefix) -> str | None:
     """
